Flatten the FIO sent request details styles
